% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights_and_outputs.R
\name{calc_weights_and_outputs}
\alias{calc_weights_and_outputs}
\title{calc_weights_and_outputs}
\usage{
calc_weights_and_outputs(
  total_portfolio,
  portfolio_type,
  output_dir,
  data_dir,
  equity_market_list,
  scenario_sources_list,
  scenario_geographies_list,
  sector_list,
  start_year,
  time_horizon
)
}
\arguments{
\item{total_portfolio}{\code{data.frame}: PACTA formatted portfolio.}

\item{portfolio_type}{\code{character}: Portfolio type. Accepts "Equity" or
"Bonds".}

\item{output_dir}{filepath: Directory to save outputs.}

\item{data_dir}{filepath: Directory with "pacta-data"}

\item{equity_market_list}{character vector: List of equity markets to
include in analysis.}

\item{scenario_sources_list}{character vector: List of scenario sources to
include in analysis. Note sources must be available in the "pacta-data".}

\item{scenario_geographies_list}{character vector: List of scenario-defined
geographies to include in analysis. Note these geographies must be available
in the scenario data.}

\item{sector_list}{character vector: List of sectors to include in analysis.}

\item{start_year}{integer: Start year for analysis.}

\item{time_horizon}{integer: Time horizon after start for analysis (usually
5 years).}
}
\value{
No return value. Saves outputs to output_dir.
}
\description{
This function calculates the weights and outputs for the
portfolio. This is the core of running the "PACTA analysis" for a portfolio.
}
