% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{run_analysis}
\usage{
run_analysis(
  pacta_data_dir,
  output_dir,
  equity_market_list,
  scenario_sources_list,
  scenario_geographies_list,
  sector_list,
  start_year,
  time_horizon
)
}
\arguments{
\item{pacta_data_dir}{filepath: Directory with "pacta-data"}

\item{output_dir}{filepath: Directory to save outputs.}

\item{equity_market_list}{character vector: List of equity markets to
include in analysis.}

\item{scenario_sources_list}{character vector: List of scenario sources to
include in analysis. Note sources must be available in the "pacta-data".}

\item{scenario_geographies_list}{character vector: List of scenario-defined
geographies to include in analysis. Note these geographies must be available
in the scenario data.}

\item{sector_list}{character vector: List of sectors to include in analysis.}

\item{start_year}{integer: Start year for analysis.}

\item{time_horizon}{integer: Time horizon after start for analysis (usually
5 years).}
}
\value{
No return value (NULL). Saves outputs to output_dir.
}
\description{
This function runs the PACTA analysis for a portfolio.
}
