% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_audit.R
\name{run_audit}
\alias{run_audit}
\title{run_audit}
\usage{
run_audit(pacta_data_dir, portfolio_dir, portfolio_files, output_dir)
}
\arguments{
\item{pacta_data_dir}{filepath: Directory with "pacta-data"}

\item{portfolio_dir}{filepath: Directory with portfolio files}

\item{portfolio_files}{character vector: List of portfolio files to process.}

\item{output_dir}{filepath: Directory to save outputs.}
}
\value{
No return value. Saves outputs to output_dir.
}
\description{
This function runs the PACTA audit for a portfolio.
}
