% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta.R
\name{run_pacta}
\alias{run_pacta}
\title{Run PACTA}
\usage{
run_pacta(
  params,
  pacta_data_dir = Sys.getenv("PACTA_DATA_DIR"),
  output_dir = Sys.getenv("ANALYSIS_OUTPUT_DIR"),
  portfolio_dir = Sys.getenv("PORTFOLIO_DIR")
)
}
\arguments{
\item{params}{List: parameters for analysis. Output of
\code{pacta.workflow.utils::parse_params}. Must contain:
\itemize{
\item portfolio: List with files key containing the name of the portfolio
file.
\item analysis: List with keys:
\itemize{
\item equityMarketList: List of equity markets to analyze.
\item scenarioGeographiesList: List of scenario geographies to analyze.
\item scenarioSourcesList: List of scenario sources to analyze.
\item sectorList: List of sectors to analyze.
\item startYear: Start year for analysis.
\item timeHorizon: Time horizon for analysis.
}
}}

\item{pacta_data_dir}{filepath: Directory with "pacta-data"}

\item{output_dir}{filepath: Directory to save outputs.}

\item{portfolio_dir}{filepath: Directory with portfolio files}
}
\value{
No return value. Saves outputs to output_dir.
}
\description{
This function runs the PACTA audit and analysis for a portfolio.
}
