% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta.R
\name{run_pacta}
\alias{run_pacta}
\title{Run PACTA}
\usage{
run_pacta(
  raw_params = commandArgs(trailingOnly = TRUE),
  pacta_data_dir = Sys.getenv("PACTA_DATA_DIR"),
  output_dir = Sys.getenv("ANALYSIS_OUTPUT_DIR"),
  portfolio_dir = Sys.getenv("PORTFOLIO_DIR")
)
}
\arguments{
\item{raw_params}{character or filepath: JSON string or file with parameters.}

\item{pacta_data_dir}{filepath: Directory with "pacta-data"}

\item{output_dir}{filepath: Directory to save outputs.}

\item{portfolio_dir}{filepath: Directory with portfolio files}
}
\value{
No return value. Saves outputs to output_dir.
}
\description{
This function runs the PACTA audit and analysis for a portfolio.
}
